+++ gmetad/conf.h
@@ -21,8 +21,8 @@ typedef struct
       char *carbon_server;
       int carbon_port;
       int carbon_timeout;
-      char *memcached_parameters;
       char *graphite_prefix;
+      char *graphite_path;
       int scalable_mode;
       int write_rrds;
       int all_trusted;
