+++ gmetad/gmetad.conf.in
@@ -154,15 +154,28 @@ case_sensitive_hostnames 0
 # default: 2003
 # carbon_port 2003
 #
-# A prefix to prepend to the metric names exported by gmetad. Graphite uses dot-
+# **Deprecated in favor of graphite_path** A prefix to prepend to the 
+# metric names exported by gmetad. Graphite uses dot-
 # separated paths to organize and refer to metrics. 
 # default: unspecified
 # graphite_prefix "datacenter1.gmetad"
 #
+# A user-definable graphite path. Graphite uses dot-
+# separated paths to organize and refer to metrics. 
+# For reverse compatibility graphite_prefix will be prepended to this
+# path, but this behavior should be considered deprecated.
+# This path may include 3 variables that will be replaced accordingly:
+# %s -> source (cluster name)
+# %h -> host (host name)
+# %m -> metric (metric name)
+# default: graphite_prefix.%s.%h.%m
+# graphite_path "datacenter1.gmetad.%s.%h.%m
+
 # Number of milliseconds gmetad will wait for a response from the graphite server 
 # default: 500
 # carbon_timeout 500
-#
+
+#-------------------------------------------------------------------------------
 # Memcached configuration (if it has been compiled in)
 # Format documentation at http://docs.libmemcached.org/libmemcached_configuration.html
 # default: ""
